+++ lib/components/input-field/input-field.jsx
@@ -89,10 +89,7 @@ export const InputField = (props) => {
 				},
 				mono: {
 					true: 'es:font-mono es:text-12!',
-					false: [
-						'es:text-13! es:font-variation-["wdth"_80,"YTLC"_520,"wght"_300,"slnt"_0]',
-						'es:tracking-wide es:placeholder-shown:font-variation-["wdth"_100,"YTLC"_500,"wght"_250,"slnt"_-8]',
-					],
+					false: 'es:text-13! es:font-variation-["wdth"_80,"wght"_325,"slnt"_0,"ROND"_100] es:placeholder-shown:font-variation-["wdth"_64,"wght"_300,"slnt"_-10,"ROND"_0]',
 				},
 				multiline: {
 					false: 'es:py-2',